Add methods that parse a URI
Instead of separate host and path args, you can now specify
the whole URI in one string.

Also fixes bugs where redirects between HTTP and HTTPS did
not work.

We would like to deprecate giving the host in the form
"hostname:portnumber" so that you cleanly choose between
a parsed input and an itemized input.  Unfortunately
there's no such thing as a runtime deprecation.

No Punycode support, so non-ASCII host name must be
pre-expanded to the xn--foo-bar notation.  It's unclear
to me whether redirects are expected to handle that, or
whether the Location: header in the redirect has
already puny-endoded host names.