Rationalize the methods on Client.
The aim is that all methods take --uri for a parseable URI,
but they also take --scheme, --host, --port, --path, and --fragment
for a pre-parsed location.

The methods that don't fit this scheme and take unnamed arguments
are marked as disappearing in the next major release, but it
hardly seems worth deprecating them now.

This is not backwards compatible for those that already started
using the web_socket methods, but it is thought that nobody has
been so fast.

Removes the error handling block for the web_socket call, which
did not fit well with the other methods and was not terribly
useful.