Make a copy of the headers in the client API.
This means internal additions to the headers do not
modify the headers object that was passed to use.

Also changes the Headers implementation to not wrap
single values in a list.

Also tries to avoid some calls to string normalization
under the assumption that map lookup is much faster than
normalization.