Support ping-pong on websockets.
This also adds a semaphore that is used in case the
user tries to write packets from more than one task
to the same websocket connection. This is needed
because when we get a ping (on the reading task) we
need to write a pong while there may be other writes
going on on another task.

Currently there is no way to do anything with pong
messages we receive, but we can send pings and ignore
the pongs that come back. And we react correctly to
pings by sending a pong.